+++ lib/src/utils/extensions.dart
@@ -123,3 +123,9 @@ extension CanvasExtension on Canvas {
 extension PathExtension on Path {
   Path scaled(double scale) => transform(Matrix4.identity().scaled(scale).storage);
 }
+
+extension ListExtension<T> on List<T> {
+  /// Returns the list with the given object inserted between each element.
+  List<T> interlacedWith(T object) =>
+      isEmpty ? [] : List.generate(length * 2 - 1, (index) => index.isEven ? this[index ~/ 2] : object);
+}